refactor: merge HttpRequestMessageExtensions
